Export senders for Meningococci
Sender export did only take Haemophilus sendings into account
hence it was missing senders that only focus on Meningococci.

Sender export will be separated same as all other exports.

#108
