+++ HaemophilusWeb/Properties/GlobalAssemblyInfo.cs
@@ -25,5 +25,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Revision and Build Numbers
 // by using the '*' as shown below:
 
-[assembly: AssemblyVersion("3.27")]
-[assembly: AssemblyFileVersion("3.27")]
+[assembly: AssemblyVersion("3.28")]
+[assembly: AssemblyFileVersion("3.28")]
