+++ HaemophilusWeb/Views/MeningoSender/Export.cshtml
@@ -2,7 +2,7 @@
 @model HaemophilusWeb.ViewModels.FromToQuery
 
 @{
-    ViewBag.Title = "Export der Einsender (Haemophilus)";
+    ViewBag.Title = "Export der Einsender (Meningokokken)";
 }
 
 <h3>Entnahmedatum der Einsendung(en)</h3>
