+++ HaemophilusWeb/Web.template.config
@@ -53,10 +53,7 @@
     <httpRuntime targetFramework="4.5" />
   </system.web>
   <system.webServer>
-    <modules>
-      <remove name="FormsAuthenticationModule" />
-    </modules>
-    <caching>
+	<caching>
       <profiles>
         <add extension=".docx" policy="DontCache" kernelCachePolicy="DontCache" />
       </profiles>
